Basic support for insteadof substitutions in repo URLs
Organization members frequently have things like this in their ~/.gitconfig:

[url http://github.com/mono/]
    insteadof = mono:

So that they can `git clone mono:subproject` instead of typing the full URL.

Unfortunately, remote.origin.url returns this shorthand instead of the full URL.

With this commit, if a user and proj can't be extracted from the origin URL
with the standard regex, one pass is made over the insteadof substitutions,
and the first matching github.com substitution is applied, and we try again
to extract user and proj.

Lastly, this commit makes the .git at the end of the URL optional, because
it is actually never required.